Custom spider settings
In order to make easier the spider configuration the settings used by the
spider will be read from the host machine. New tasks has been added in the
script to configure the settings.py file.  This file is generated in the
configuration directory defined in the variables and mapped as a read-only
volume in the container used to run the spider.

Signed-off-by: José Guilherme Vanz <jvanz@jvanz.com>